+++ firmware/controllers/sensors/converters/resistance_func.cpp
@@ -3,7 +3,6 @@
  */
 
 #include "resistance_func.h"
-#include "loggingcentral.h"
 
 void ResistanceFunc::configure(float supplyVoltage, float pullupResistor) {
 	m_pullupResistor = pullupResistor;
@@ -25,9 +24,4 @@ SensorResult ResistanceFunc::convert(float raw) const {
 	float resistance = m_pullupResistor / (m_supplyVoltage / raw - 1);
 
 	return resistance;
-}
-
-void ResistanceFunc::showInfo(Logging* logger, float testInputValue) const {
-	const auto result = convert(testInputValue);
-	scheduleMsg(logger, "    %.2f volts -> %.1f ohms with supply voltage %.2f and pullup %.1f.", testInputValue, result.Value, m_supplyVoltage, m_pullupResistor);
 }
