+++ firmware/controllers/sensors/sensor_info_printing.cpp
@@ -2,6 +2,9 @@
 #include "proxy_sensor.h"
 #include "functional_sensor.h"
 #include "redundant_sensor.h"
+#include "linear_func.h"
+#include "resistance_func.h"
+#include "thermistor_func.h"
 #include "efilib.h"
 #include "loggingcentral.h"
 
@@ -29,5 +32,21 @@ void CanSensorBase::showInfo(Logging* logger, const char* sensorName) const {
 #endif // EFI_CAN_SUPPORT
 
 void RedundantSensor::showInfo(Logging* logger, const char* sensorName) const {
-	scheduleMsg(logger, "Redundant sensor \"%s\" combining \"%s\" and \"%s\"", sensorName, getSensorName(m_first), getSensorName(m_second));
+	scheduleMsg(logger, "Sensor \"%s\" is redundant combining \"%s\" and \"%s\"", sensorName, getSensorName(m_first), getSensorName(m_second));
+}
+
+void LinearFunc::showInfo(Logging* logger, float testRawValue) const {
+	scheduleMsg(logger, "    Linear function slope: %.2f offset: %.2f min: %.1f max: %.1f", m_a, m_b, m_minOutput, m_maxOutput);
+	const auto [valid, value] = convert(testRawValue);
+	scheduleMsg(logger, "      raw value %.2f converts to %.2f valid: %d", testRawValue, value, valid);
+}
+
+void ResistanceFunc::showInfo(Logging* logger, float testInputValue) const {
+	const auto result = convert(testInputValue);
+	scheduleMsg(logger, "    %.2f volts -> %.1f ohms with supply voltage %.2f and pullup %.1f.", testInputValue, result.Value, m_supplyVoltage, m_pullupResistor);
+}
+
+void ThermistorFunc::showInfo(Logging* logger, float testInputValue) const {
+	const auto [valid, value] = convert(testInputValue);
+	scheduleMsg(logger, "    %.1f ohms -> valid: %d. %.1f deg C", testInputValue, valid, value);
 }
